requirements.txt
anthropic>=0.30
fastapi>=0.111
uvicorn>=0.30
python-dotenv>=1.0
pytest>=8.0
pydantic~=2.11.7
requests~=2.32.4